feat: implement getMbr on mask
Refs: #121
