+++ src/maskAnalysis/getMbr.ts
@@ -1,0 +1,95 @@
+import { Mask } from '../Mask';
+import { difference, normalize, Point, rotate } from '../utils/geometry/points';
+
+import { getExtendedBorderPoints } from './utils/getExtendedBorderPoints';
+import { monotoneChainConvexHull } from './utils/monotoneChainConvexHull';
+
+/**
+ * Get the four corners of the minimun bounding rectangle of an ROI.
+ *
+ * @param mask - The ROI to process.
+ * @returns The array of corners.
+ */
+export function getMbr(mask: Mask): Point[] {
+  const vertices = monotoneChainConvexHull(getExtendedBorderPoints(mask));
+
+  if (mask.size === 1 && mask.getBit(0, 0) === 1) {
+    return [vertices[0], vertices[0], vertices[0], vertices[0]];
+  }
+
+  let rotatedVertices: Point[] = [];
+  let minSurface = Number.POSITIVE_INFINITY;
+  let minSurfaceAngle = 0;
+  let mbr: Point[] = [];
+
+  for (let i = 0; i < vertices.length; i++) {
+    let angle = getAngle(vertices[i], vertices[(i + 1) % vertices.length]);
+
+    rotatedVertices = rotate(-angle, vertices);
+    // console.log({ rotatedVertices });
+
+    // we rotate and translate so that this segment is at the bottom
+    let aX = rotatedVertices[i].column;
+    let aY = rotatedVertices[i].row;
+    let bX = rotatedVertices[(i + 1) % rotatedVertices.length].column;
+    let bY = rotatedVertices[(i + 1) % rotatedVertices.length].row;
+
+    let tUndefined = true;
+    let tMin = 0;
+    let tMax = 0;
+    let maxWidth = 0;
+    for (let point of rotatedVertices) {
+      let cX = point.column;
+      let cY = point.row;
+      let t = (cX - aX) / (bX - aX);
+      if (tUndefined === true) {
+        tUndefined = false;
+        tMin = t;
+        tMax = t;
+      } else {
+        if (t < tMin) tMin = t;
+        if (t > tMax) tMax = t;
+      }
+      let width = (-(bX - aX) * cY + bX * aY - bY * aX) / (bX - aX);
+
+      if (Math.abs(width) > Math.abs(maxWidth)) {
+        maxWidth = width;
+      }
+    }
+    let minPoint = { column: aX + tMin * (bX - aX), row: aY };
+    let maxPoint = { column: aX + tMax * (bX - aX), row: aY };
+
+    let currentSurface = Math.abs(maxWidth * (tMin - tMax) * (bX - aX));
+
+    if (currentSurface < minSurface) {
+      minSurfaceAngle = angle;
+      minSurface = currentSurface;
+      mbr = [
+        maxPoint,
+        minPoint,
+        { column: minPoint.column, row: minPoint.row - maxWidth },
+        { column: maxPoint.column, row: maxPoint.row - maxWidth },
+      ];
+    }
+  }
+
+  const mbrRotated = rotate(minSurfaceAngle, mbr);
+
+  return mbrRotated;
+}
+
+/**
+ *  The angle that allows to make the line going through p1 and p2 horizontal.
+ *  This is an optimized version because it assumes that one of the vectors is horizontal.
+ *
+ * @param p1 - First point.
+ * @param p2 - Second point.
+ * @returns Rotation angle to make the line horizontal.
+ */
+function getAngle(p1: Point, p2: Point): number {
+  let diff = difference(p2, p1);
+  let vector = normalize(diff);
+  let angle = Math.acos(vector.column);
+  if (vector.row < 0) return -angle;
+  return angle;
+}
